+++ packages/categories/client/Router.js
@@ -51,6 +51,33 @@ Router.map(function() {
 	});
 
 	this.route('panelCats', {
+		path: '/categories',
+		template: 'panelList',
+		layoutTemplate: 'layoutMap',
+		loadingTemplate: 'pageLoading',
+		onBeforeAction: function () {
+			if(!K.Admin || !K.Admin.isMe())
+				Router.go('root');
+			else
+				this.next();
+		},
+		waitOn: function() {
+			Session.set('showSidebar', true);
+			return Meteor.subscribe('catsByName');
+		},
+		data: function() {
+			if(!this.ready()) return null;
+			return  {
+				title: i18n('label_cats'),
+				className: 'cats-name',
+				headerTemplate: 'formSearchCats',
+				itemsTemplate: 'itemCat',
+				items: K.findCatsByName('').fetch()
+			};
+		}
+	});
+
+	this.route('panelCatsType', {
 		path: '/categories/:type',
 		template: 'panelList',
 		layoutTemplate: 'layoutMap',